pyproject.toml
[project]
name = "sre-bot"
version = "0.1.0"
description = "SRE Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "atlassian-python-api>=4.0.4",
    "jira>=3.8.0",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.2",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "langgraph-checkpoint-postgres>=2.0.21",
    "lxml>=5.4.0",
    "pydantic>=2.11.4",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]
